requirements.txt
certifi==2020.12.5
chardet==4.0.0
csv342==1.0.1
Cython==0.29.21
idna==2.10
Pillow==8.1.0
pkg-resources==0.0.0
pytesseract==0.3.7
requests==2.25.1
tesseract==0.1.3
tesseract-ocr==0.0.1
urllib3==1.26.3
